+++ astropy/tests/plugins/remotedata.py
@@ -1,13 +1,13 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 """
-These plugins modify the behavior of py.test and are meant to be imported
-into conftest.py in the root directory.
+This plugin provides command-line options for controlling whether and how tests
+make use of online data.
 """
 from __future__ import (absolute_import, division, print_function,
                         unicode_literals)
 
-from .helper import pytest
-from .disable_internet import turn_off_internet, turn_on_internet
+from ..helper import pytest
+from ..disable_internet import turn_off_internet, turn_on_internet
 
 
 def pytest_addoption(parser):